+++ libs/light/create.ts
@@ -20,4 +20,25 @@ namespace light {
     ): NeoPixelStrip {
         return light.createNeoPixelStrip(pin, numleds, mode);
     }
+
+    /**
+     * Creates a strip of colored LEDs (APA102)
+     */
+    //% blockId="light_create_dotstar" block="create APA102 strip|data %data|clock %clk|with %numleds pixels"
+    //% help="light/create-apa102-strip"
+    //% trackArgs=0,1,2
+    //% parts="dotstar"
+    //% weight=100 blockSetVariable=strip
+    //% advanced=true
+    export function createAPA102Strip(
+        dataPin: DigitalInOutPin,
+        clkPin: DigitalInOutPin,
+        numleds: number): NeoPixelStrip {
+        const strip = new NeoPixelStrip();
+        strip._mode = NeoPixelMode.APA102;
+        strip._length = Math.max(0, numleds | 0);
+        strip._dataPin = dataPin;
+        strip._clkPin = clkPin;
+        return strip;
+    }
 }
