+++ libs/light/pxt.json
@@ -8,6 +8,7 @@
         "defaultstrip.ts",
         "onboardstrip.ts",
         "defaultlights.ts",
+        "defaultlightsoverrides.ts",
         "ns.ts",
         "jacdac.ts",
         "neopixeloverrides.ts",
